+++ client/packages/host/src/routers/CatalogueRouter.tsx
@@ -14,10 +14,6 @@ const MasterListService = React.lazy(
   () => import('@openmsupply-client/system/src/MasterList/Service/Service')
 );
 
-const ImmunisationService = React.lazy(
-  () => import('@openmsupply-client/system/src/Immunisation/Service/Service')
-);
-
 const fullAssetPath = RouteBuilder.create(AppRoute.Catalogue)
   .addPart(AppRoute.Assets)
   .addWildCard()
@@ -30,11 +26,6 @@ const fullItemPath = RouteBuilder.create(AppRoute.Catalogue)
 
 const fullMasterListPath = RouteBuilder.create(AppRoute.Catalogue)
   .addPart(AppRoute.MasterLists)
-  .addWildCard()
-  .build();
-
-const immunisationFullPath = RouteBuilder.create(AppRoute.Catalogue)
-  .addPart(AppRoute.Immunisations)
   .addWildCard()
   .build();
 
@@ -52,10 +43,6 @@ export const CatalogueRouter: FC = () => {
 
   if (gotoMasterLists) {
     return <MasterListService />;
-  }
-
-  if (immunisationFullPath) {
-    return <ImmunisationService />;
   }
 
   const notFoundRoute = RouteBuilder.create(AppRoute.PageNotFound).build();
