+++ client/packages/host/src/routers/index.ts
@@ -6,3 +6,4 @@ export * from './ReplenishmentRouter';
 export * from './DispensaryRouter';
 export * from './ColdChainRouter';
 export * from './ManageRouter';
+export * from './ProgramsRouter';
